Add a new field to CRDs: template
It's a successor of `secretsTemplates` feature, that should replace it in
the new API version. It's supposed to be an easily extendable and
flexible solution for generating different variable that an application
may need.

The main idea is that instead of populating an object with predefined
values that can be used for rendering a template, it should be able to
access any data from all the main sources:
  - Secret
  - ConfigMap
  - DatabaseCR
  - DbUserCR
  - DbInstanceCR
  - Database queries (experimental)

And it would mean that instead of extending template functionality, we
can deliver operator features and templates should get them
automatically

Issue: https://github.com/db-operator/db-operator/issues/47