Remove the kci copyright from the generated code
The code is generated, so I don't understand why it should be
copyrighted. But it's not the main reason wh I've done it. The problem
is that we hused to have the kci copyright in the boilreplate with the
license. And since we should not use it anymore, it was removed from
there. But all the generated files that generation is triggered by make,
recieved a new copyright. So it seems like a lot of manual work to
preserve it, and it doesn't seems right. All the hand-written files
still have the copyright.